# backend/openapi.yaml
openapi: 3.1.0
info:
  title: itgel API
  version: 1.0.0
  description: |
    Захиалгын дэлгүүрийн backend API.

    - Бүх мөнгөн дүн бүхэл тоо (₮).
    - Бүх огноо ISO 8601, `Asia/Ulaanbaatar` (UTC+8) цагаар бодогдоно.
    - Хэрэглэгчийн API `costPrice`-г хэзээ ч буцаахгүй.
servers:
  - url: http://localhost:4000
    description: Dev
...
security: []
paths:
...
  /api/products:
    get:
      tags: [public]
      summary: Барааны жагсаалт
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [order, ready] } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 60 } }
        - { name: sort, in: query, schema: { type: string, enum: [new, priceAsc, priceDesc, closing] } }
      responses:
        '200':
          description: Хуудаслалттай жагсаалт
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PublicProduct' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
  /api/products/{id}:
    get:
      tags: [public]
      summary: Барааны дэлгэрэнгүй (variants, sizeChart, arriveFrom/To)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Бараа
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/PublicProduct' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/products:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Барааны жагсаалт (costPrice, ашигтай)
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ProductStatus' } }
        - { name: category, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [order, ready] } }
        - { name: q, in: query, schema: { type: string } }
        - { name: includeDeleted, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Жагсаалт
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AdminProduct' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Бараа үүсгэх
      description: leadMinDays/leadMaxDays илгээгээгүй бол тохиргооны анхны утга (7–14) орно.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '201': { description: Үүссэн бараа }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/admin/products/{id}:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Барааны дэлгэрэнгүй
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Бараа }
    patch:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Бараа засах
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '200': { description: Засагдсан бараа }
    delete:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Бараа устгах (soft delete)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Устгасан тоо }
...
  /api/admin/products/{id}/images:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Зураг байршуулах presigned URL
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contentType]
              properties:
                contentType: { type: string, example: image/jpeg }
                fileName: { type: string }
      responses:
        '200':
          description: Presigned upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      uploadUrl: { type: string }
                      publicUrl: { type: string }
                      key: { type: string }
                      method: { type: string, enum: [PUT] }
                      expiresInSec: { type: integer }
                      provider: { type: string, enum: [supabase, r2, mock] }
    patch:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Байршуулсан зургийн URL-уудыг бүртгэх
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [images]
              properties:
                images: { type: array, items: { type: string, format: uri } }
      responses:
        '200': { description: Зургийн жагсаалт }
...
  /api/admin/reports/products:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Бараа бүрийн зарагдсан / буцаасан тоо, дүн
      parameters:
        - { name: period, in: query, schema: { type: string, enum: ['3m', '6m', '1y'] } }
        - { name: productIds, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 200 } }
      responses:
        '200': { description: Жагсаалт }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Манай JWT (`/api/auth/verify` эсвэл `/api/admin/auth/login`) эсвэл
        Supabase Auth-ийн access token (JWKS-ээр шалгагдана).
  responses:
    BadRequest:
      description: Оруулсан утга буруу
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Олдсонгүй
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        pages: { type: integer }
    ProductStatus:
      type: string
      enum: [ACTIVE, HIDDEN, DRAFT, CLOSED, SOLD_OUT, ARCHIVED]
...
    PublicProduct:
      type: object
      description: costPrice энд хэзээ ч байхгүй.
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        categoryId: { type: string }
        price: { type: integer }
        stock: { type: integer }
        type: { type: string, enum: [order, ready] }
        status: { $ref: '#/components/schemas/ProductStatus' }
        closeAt: { type: string, format: date-time, nullable: true }
        leadMinDays: { type: integer }
        leadMaxDays: { type: integer }
        arriveFrom: { type: string, format: date-time }
        arriveTo: { type: string, format: date-time }
        images: { type: array, items: { type: string } }
        sizes: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        optionPrices:
          type: array
          items:
            type: object
            properties:
              kind: { type: string }
              value: { type: string }
              selections:
                type: object
                additionalProperties: { type: string }
              price: { type: integer }
        skuStocks:
          type: array
          items:
            type: object
            properties:
              selections:
                type: object
                additionalProperties: { type: string }
              stock: { type: integer }
        sizeChart:
          type: array
          items:
            type: object
            properties:
              size: { type: string }
              heightRange: { type: string }
              chestCm: { type: string }
    AdminProduct:
      allOf:
        - $ref: '#/components/schemas/PublicProduct'
        - type: object
          properties:
            costPrice: { type: integer }
            sellPrice: { type: integer }
            profit: { type: integer }
            marginPercent: { type: integer }
            deletedAt: { type: string, format: date-time, nullable: true }
    ProductInput:
      type: object
      required: [name, categoryId, costPrice, sellPrice]
      properties:
        name: { type: string }
        description: { type: string }
        categoryId: { type: string }
        costPrice: { type: integer }
        sellPrice: { type: integer }
        stock: { type: integer }
        closeAt: { type: string, format: date-time, nullable: true }
        leadMinDays: { type: integer }
        leadMaxDays: { type: integer }
        status: { $ref: '#/components/schemas/ProductStatus' }
        images: { type: array, items: { type: string, format: uri } }
        sizes: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        sizeChart:
          type: array
          items:
            type: object
            required: [size]
            properties:
              size: { type: string }
              heightRange: { type: string }
              chestCm: { type: string }
...